Prevent queries for the null customer
If a customer ID is programmatically retrieved, it is possible that the
null ID is returned. The code as currently written happily formats
"null" into the API URL and attempts to fetch devices for it, which
always results in a fetch error. This change sends the user an error
message and prevents accidentally querying for the null customer.